add file exists check

+++ lib/github_collaborators/terraform_block_creator.rb
@@ -82,17 +82,25 @@ class GithubCollaborators
                 # For each repository 
                 repositories.each { |repo|
 
-                    # Read the relevant file into array
-                    file = File.read(repo_file(repo))
-                            .split("\n")
+                    if File.exists(repo_file(repo))
 
-                    # Create insert contents and insert, length - 2 gives the correct location in the file
-                    file.insert((file.length) - 2, create_insert)
+                        # Read the relevant file into array
+                        file = File.read(repo_file(repo))
+                                .split("\n")
 
-                    # Write file
-                    File.open(repo_file(repo), "w") { 
-                        |f| f.puts(file)  
-                    }
+                        # Create insert contents and insert, length - 2 gives the correct location in the file
+                        file.insert((file.length) - 2, create_insert)
+
+                        # Write file
+                        File.open(repo_file(repo), "w") { 
+                            |f| f.puts(file)  
+                        }
+                    else
+                        # Write to blank file
+                        File.open(repo_file(repo), "w") { 
+                            |f| f.puts(file)  
+                        }
+                    end
                 }
             end
         end
